[Apple] Fix build
We weren't explicitly including `bzip` for RSL, so the `wbz` crate was actually covering our missing dep on MacOS (via the `bz2` crate). So when we switched to the Rust version of `bz2`, we no longer had the duplicate dependency, hence the build broke.

Now, we update `zip` and fix a similar bug with `liblzma` (XZ). (New `zip` dep?) Might need to adjust this on Linux, too.

(These CMake scripts should be refactored considerably)